requirements.txt
# Core Framework
streamlit>=1.37.0
python-dotenv>=1.0.0

# LLM API (FREE!)
groq>=0.4.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# MLOps (Future phases)
# mlflow>=2.10.0
# prometheus-client>=0.19.0
